Added -truncatetext and -growtext options to textbox and textfield.
-truncatetext truncates the text to the current maximum size.  This
should always work and never return an exception.

-growtext attempts to set the maximum size to accomodate the text
argument.  This could potentially fail if the device can't accommodate
a maxsize that large.  Since these are 'sensible default' type of
options and not 'fine grained control' tools, I think that in that
case, rather than fail, we'll just accept the maximum size it gives us
and use that, even if it ends up truncating some text.